+++ src/main/java/it/unibo/sampleapp/model/object/impl/Watergirl.java
@@ -16,7 +16,7 @@ public class Watergirl extends AbstractPlayer {
      * @param width player width
      * @param height player height
      */
-    public Watergirl(final int startX, final int startY, final int width, final int height) {
+    public Watergirl(final double startX, final double startY, final int width, final int height) {
         super(startX, startY, width, height);
     }
